adv: bring-up latest android versions (1/2)

+++ apply-patches.sh
@@ -34,27 +34,6 @@
 #          #
 ############
 
-BOLD="\033[1m"
-GREEN="\033[01;32m"
-RED="\033[01;31m"
-RESTORE="\033[0m"
-
-
-###############
-#             #
-#  FUNCTIONS  #
-#             #
-###############
-
-# PRINTS A FORMATTED HEADER TO POINT OUT WHAT IS BEING DONE TO THE USER
-function echoText() {
-    echo -e ${RED}
-    echo -e "====$( for i in $( seq ${#1} ); do echo -e "=\c"; done )===="
-    echo -e "==  ${1}  =="
-    echo -e "====$( for i in $( seq ${#1} ); do echo -e "=\c"; done )===="
-    echo -e ${RESTORE}
-}
-
 # FORMATS THE TIME
 function format_time() {
     MINS=$(((${1}-${2})/60))
@@ -91,12 +70,13 @@ function format_time() {
 
 # PRINTS A HELP MENU
 function help_menu() {
-    echo -e "\n${BOLD}OVERVIEW:${RESTORE} Merges full GSI changes from phhusson into a ROM set of repos\n"
-    echo -e "${BOLD}USAGE:${RESTORE} bash apply_patches.sh <source_dir> <branch_name>\n"
-    echo -e "${BOLD}EXAMPLE:${RESTORE} bash apply_patches.sh ~/Android/Lineage android-10.0-r14\n"
-    echo -e "${BOLD}Required options:${RESTORE}"
+    echo -e "\n${BLD}OVERVIEW:${RST} Merges full GSI changes from phhusson into a ROM set of repos\n"
+    echo -e "${BLD}USAGE:${RST} bash apply_patches.sh <source_dir> <branch_name>\n"
+    echo -e "${BLD}EXAMPLE:${RST} bash apply_patches.sh ~/Android/Lineage android-10.0\n"
+    echo -e "${BLD}Required options:${RST}"
     echo -e "       source_dir: Location of the ROM tree; this needs to exist for the script to properly proceed\n"
-    echo -e "       branch_name: what commits need be pulled. Oreo or pie? \n"
+    echo -e "       branch_name: Android version, a string that starts with \"android-\" and followed by version:"
+    echo -e "       8.1, 9.0, 10.0, 11.0, 12.1\n"
 }
 
 # CREATES A NEW LINE IN TERMINAL
@@ -104,38 +84,18 @@ function newLine() {
     echo -e ""
 }
 
-# PRINTS AN ERROR IN BOLD RED
-function reportError() {
-    RED="\033[01;31m"
-    RESTORE="\033[0m"
-
-    echo -e ""
-    echo -e ${RED}"${1}"${RESTORE}
-    if [[ -z ${2} ]]; then
-        echo -e ""
-    fi
-}
-
-# PRINTS AN WARNING IN BOLD YELLOW
-function reportWarning() {
-    YELLOW="\033[01;33m"
-    RESTORE="\033[0m"
-
-    echo -e ""
-    echo -e ${YELLOW}"${1}"${RESTORE}
-    if [[ -z ${2} ]]; then
-        echo -e ""
-    fi
-}
-
 ###############
 #             #
 #  VARIABLES  #
 #             #
 ###############
+BASE_REPOS="external/selinux frameworks/base frameworks/native frameworks/av frameworks/opt/telephony system/bt system/core system/netd system/vold"
+COMMON_REPOS="bionic packages/apps/Settings"
+EXTRA_REPOS="bootable/recovery"
+NEW_REPOS="packages/apps/Bluetooth system/extras system/linkerconfig system/nfc"
 
 if [[ ! $# -ge 2 ]]; then
-    reportError "Source directory/Branch not specified!" -c; help_menu && exit
+    reportError "Source directory/Branch not specified!"
 fi
 
 while [[ $# -ge 2 ]]; do
@@ -144,75 +104,21 @@ while [[ $# -ge 2 ]]; do
             help_menu && exit ;;
         *)
             SOURCE_DIR=${1}
+            BRANCH_VERSION=${2}
             if [[ ! -d ${SOURCE_DIR} ]]; then
-                reportError "Source directory not found!" && exit
+                reportError "Source directory not found!"
             elif [[ ! -d ${SOURCE_DIR}/.repo ]]; then
-                reportError "This is not a valid Android source folder as there is no .repo folder!" && exit
+                reportError "This is not a valid Android source folder as there is no .repo folder!"
             fi ;;
     esac
 
-    # variants
-    PIE=false
-    Q=false
-    
-    case "$2" in
-        android-8.1)
-            BRANCH=android-8.1.0_r65-phh
-SUBS_REPOS="
-build
-external/selinux
-frameworks/av
-frameworks/base
-frameworks/native
-frameworks/opt/telephony
-system/bt
-system/core
-system/libvintf
-system/netd
-system/vold"
-;;
-        android-9.0)
-            PIE=true
-            BRANCH=android-9.0.0_r47-phh
-SUBS_REPOS="
-build
-external/selinux
-frameworks/av
-frameworks/base
-frameworks/native
-frameworks/opt/net/wifi
-frameworks/opt/telephony
-packages/apps/Settings
-packages/services/Telephony
-system/bt
-system/core
-system/netd
-system/sepolicy
-system/vold"
-;;
-    android-10.0)
-       Q=true
-       BRANCH=android-10.0.0_r31-phh
-SUBS_REPOS="
-bionic
-bootable/recovery
-build
-external/selinux
-external/skia
-frameworks/av
-frameworks/base
-frameworks/native
-frameworks/opt/net/wifi
-frameworks/opt/telephony
-packages/apps/Settings
-packages/services/Telephony
-system/bpf
-system/bt
-system/core
-system/netd
-system/sepolicy
-system/vold"
-;;
+    case "${BRANCH_VERSION}" in
+        android-8.1) REPOS="build system/libvintf system/sepolicy ${BASE_REPOS}";;
+        android-9.0) REPOS="build frameworks/opt/net/wifi packages/services/Telephony system/sepolicy system/libvintf ${BASE_REPOS} ${COMMON_REPOS}";;
+        android-10.0) REPOS="build packages/services/Telephony external/skia frameworks/opt/net/wifi system/bpf system/sepolicy ${BASE_REPOS} ${COMMON_REPOS} ${EXTRA_REPOS}";;
+        android-11.0) REPOS="external/skia frameworks/opt/net/wifi packages/services/Telephony system/memory/lmkd ${BASE_REPOS} ${COMMON_REPOS} ${EXTRA_REPOS} ${NEW_REPOS}";;
+        android-12.1) REPOS="frameworks/opt/net/ims hardware/interfaces packages/modules/Wifi system/bpf system/security ${BASE_REPOS} ${COMMON_REPOS} ${EXTRA_REPOS} ${NEW_REPOS}";;
+        *) reportError "Unknown android tag ($2)";;
     esac
 
     shift
@@ -228,64 +134,89 @@ unset RESULT_STRING
 
 # START TRACKING TIME
 START=$( date +%s )
+# SET BRANCH NAME (might break in future android sources :0)
+BRANCH=$( git ls-remote https://github.com/phhusson/platform_frameworks_base | awk -F'/' "(match(\$3, /^${BRANCH_VERSION}.0_r([0-9]+)(-r[0-9]+)?-phh$/, r) && r[1]>m) { m=r[1]; l=\$3 } END { print l }" )
+# ABORT IF NO BRANCHE WAS MATCHED
+[ -z "$BRANCH" ] && reportError "Failed to fetch upstream branch!"
 
-for FOLDER in ${SUBS_REPOS}; do
+echo -e "Targeting: \n\t  ${YLW}$(echo ${REPOS} | sed -e 's/\(\w\+\)\s/\1\n\t  /g') ${RST}\nfrom branch: \n\t  ${GRN}${BRANCH}${RST}"
+for FOLDER in ${REPOS}; do
+    # GIT LOG QUERY OPTIONS
+    GIT_OPTIONS="--format=%H --committer=\"Pierre-Hugues\" --invert-grep -E --grep=\"^(Merge tag|Update SPL)\" FETCH_HEAD"
     # PRINT TO THE USER WHAT WE ARE DOING
-    newLine; echoText "Merging ${FOLDER}"
+    newLine; printText "Merging ${FOLDER}"
     
     # SHIFT TO PROPER FOLDER
     cd ${SOURCE_DIR}
-   
+
     if [[ ${FOLDER} = "build" ]]; then
         # build is build/make
-	cd ${FOLDER}/make
+	echo cd ${FOLDER}/make
     else
-        cd ${FOLDER}
+        echo cd ${FOLDER}
     fi
 
     # SET PROPER URL
     URL=platform_$( echo ${FOLDER} | sed "s/\//_/g" )
 
     # FETCH THE REPO
-    git fetch https://github.com/phhusson/${URL} ${BRANCH}
+    newLine; echo "Fetching phh HEAD..."
+    git fetch https://github.com/phhusson/${URL} ${BRANCH} >/dev/null 2>&1 || reportError "! Failed to fetch ${URL} branch ${BRANCH}"
 
     # GIT GYMNASTICS (GETS MESSY, BEWARE)
-    # FIRST HASH WILL ALWAYS BE THE FETCH HEAD
-    FIRST_HASH=$( git log --format=%H -1 FETCH_HEAD)
+    # HEAD HASH WILL ALWAYS BE THE FETCH HEAD (LATEST COMMIT)
+    HEAD_HASH=$( eval git log -1 ${GIT_OPTIONS} )
 
-    # SECOND HASH WILL BE THE LAST THING I COMMITTED
-    NUMBER_OF_COMMITS=$(( $( git log --format=%H --committer="Pierre-Hugues" FETCH_HEAD | wc -l ) - 1 ))
-    SECOND_HASH=$( git log --format=%H --committer="Pierre-Hugues" FETCH_HEAD~${NUMBER_OF_COMMITS}^..FETCH_HEAD~${NUMBER_OF_COMMITS} )
-
-    # RESET ANY LOCAL CHANGES SO THAT CHERRY-PICK DOES NOT FAIL
+    # TAIL HASH WILL BE THE **FIRST** THING PHH COMMITTED
+    # NO NEED TO MINUS 1
+    NUMBER_OF_COMMITS=$( eval git log ${GIT_OPTIONS} | wc -l )
+    
+    if [[ "${NUMBER_OF_COMMITS}" -eq 0 ]]; then
+         newLine; reportWarning "SKKIPED ${FOLDER}"
+         RESULT_STRING+="${FOLDER}: ${YLW}SKIPPED${RST}\n"
+         continue
+    fi
+  
+    # NOW, SET TAIL_HASH WITH POSITIVE ${NUMBER_OF_COMMITS}
+    TAIL_HASH=$( eval git log ${GIT_OPTIONS}~${NUMBER_OF_COMMITS}^..FETCH_HEAD~${NUMBER_OF_COMMITS} )
+    
+    # TAIL HASH CRYING?
+    [ -z "${TAIL_HASH}" ] && TAIL_HASH=$( eval git log ${GIT_OPTIONS} | tail -1 )
+    
+    # RESET ANY LOCAL CHANGES
     git reset --hard HEAD
 
-    # PICK THE COMMITS IF EVERYTHING CHECKS OUT
-    if [[ ${FOLDER} = "system/vold" ]]; then
-	    ($PIE && git cherry-pick --allow-empty-message --keep-redundant-commits -X thiers 13a34a80c433dd2a5a2c195b3c568990ef9908fd^..${FIRST_HASH})
-	    ($Q && git cherry-pick --allow-empty-message --keep-redundant-commits -X thiers 037344ea364f21a80c55240f419ff11251a90c0e^..${FIRST_HASH})
-    else
-	    git cherry-pick --allow-empty-message --keep-redundant-commits -X thiers ${SECOND_HASH}^..${FIRST_HASH}
-    fi
+    LAST_COMMIT_MSG=$(git show -s --format=%s $HEAD_HASH)
+    FIRST_COMMIT_MSG=$(git show -s --format=%s $TAIL_HASH)
+    echo -e "${GRN}Applying ${NUMBER_OF_COMMITS} patch(es)${RST}:\n from: ${FIRST_COMMIT_MSG}\n to:   ${LAST_COMMIT_MSG}"
+
+    # PICK THE COMMITS
+    CHERRY_PICK="${TAIL_HASH}^..${HEAD_HASH}"
+    [  "${TAIL_HASH}" == "${HEAD_HASH}" ] && CHERRY_PICK="${TAIL_HASH}"
+    eval git cherry-pick --allow-empty-message --keep-redundant-commits -X thiers ${CHERRY_PICK}
 
     # ADD TO RESULT STRING
     if [[ $? -ne 0 ]]; then
-        RESULT_STRING+="${FOLDER}: ${RED}FAILED${RESTORE}\n"
+        RESULT_STRING+="${FOLDER}: ${RED}FAILED${RST}\n"
+        CLR="${RED}" 
     else
-        RESULT_STRING+="${FOLDER}: ${GREEN}SUCCESS${RESTORE}\n"
+        RESULT_STRING+="${FOLDER}: ${GRN}SUCCESS${RST}\n"
+        CLR="${GRN}"
     fi
+    
+    newLine; printText "DONE ${FOLDER}" ${CLR}
 done
 
 # SHIFT BACK TO THE TOP OF THE REPO
 cd ${SOURCE_DIR}
 
 # PRINT RESULTS
-echoText "RESULTS"
+printText "RESULTS"
 echo -e ${RESULT_STRING}
 
 # STOP TRACKING TIME
 END=$( date +%s )
 
 # PRINT RESULT TO USER
-echoText "SCRIPT COMPLETED!"
-echo -e ${RED}"TIME: $(format_time ${END} ${START})"${RESTORE}; newLine
+printText "SCRIPT COMPLETED!"
+echo -e ${RED}"TIME: $(format_time ${END} ${START})"${RST}; newLine
